pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "polyfund"
version = "0.1.0"
description = "Polymarket thematic quant basket builder"
readme = "AGENTS.md"
requires-python = ">=3.9"
license = {text = "UNLICENSED"}
authors = [{name = "Polyfund"}]

dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "rapidfuzz>=3.0",
  "requests>=2.31",
  "tenacity>=8.2",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "supabase>=2.4",
  "google-genai>=0.5",
  "python-dotenv>=1.2.1",
]

[tool.setuptools]
packages = ["backend"]
